+++ libraries/core-react/src/Tooltip/Tooltip.test.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef */
 import React from 'react'
-import { render, cleanup } from '@testing-library/react'
+import { render, cleanup, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
 import 'jest-styled-components'
 import styled from 'styled-components'
@@ -21,31 +21,39 @@ afterEach(cleanup)
 
 describe('Tooltip', () => {
   it('Tooltip has correct placement', () => {
-    const { container } = render(<Tooltip placement="topRight">Anchor</Tooltip>)
-    const tooltipWrapper = container.lastChild
-    const tooltip = tooltipWrapper.firstChild
+    const { container } = render(
+      <Tooltip open placement="topRight">
+        Anchor
+      </Tooltip>,
+    )
+    const tooltipWrapper = container.firstChild
+    const tooltip = screen.getByRole('tooltip')
 
     expect(tooltipWrapper).toHaveStyleRule('display', 'flex')
     expect(tooltip).toHaveStyleRule('top', `${topRight.tooltipTop}`)
     expect(tooltip).toHaveStyleRule('right', `${topRight.tooltipRight}`)
   })
   it('Arrow has correct placement', () => {
-    const { container } = render(<Tooltip placement="topRight">Anchor</Tooltip>)
+    const { container } = render(
+      <Tooltip open placement="topRight">
+        Anchor
+      </Tooltip>,
+    )
     const arrow = container.lastElementChild.lastChild.firstChild.firstChild
     expect(arrow).toHaveStyleRule('right', `${topRight.arrowRight}`)
     expect(arrow).toHaveStyleRule('bottom', `${topRight.arrowBottom}`)
   })
   it('Has provided necessary props', () => {
     const title = 'Title'
-    const variant = 'large'
+    const placement = 'topLeft'
     const anchor = 'Anchor'
     const { queryByText } = render(
-      <Tooltip variant={variant} title={title}>
+      <Tooltip open placement={placement} title={title}>
         {anchor}
       </Tooltip>,
     )
     expect(queryByText(title)).toBeDefined()
-    expect(queryByText(variant)).toBeDefined()
+    expect(queryByText(placement)).toBeDefined()
     expect(queryByText(anchor)).toBeDefined()
   })
   it('Can extend the css for the component', () => {
