+++ libraries/core-react/src/Tooltip/index.js
@@ -1,2 +1,1 @@
 export { Tooltip } from './Tooltip'
-export { TooltipItem } from './TooltipItem'
